requirements.txt
playwright==1.46.0
gspread==6.1.2
google-auth==2.34.0
requests==2.32.3
python-dotenv==1.0.1
pytest==8.3.2
